Fix: 에러수정

+++ srcs/requirements/user_server/django/oauth/views/otp.py
@@ -12,40 +12,29 @@ from rest_framework import status
 
 class OTPCheckView(APIView):
 	def post(self, request):
-		# now = timezone.now()
-		# pk = request.COOKIES.get('pk')
-		# otp = OTPModel.objects.get(user=pk)
-  
-		# form_data = json.loads(request.body)
-
-		# if now - otp.created_at > timedelta(minutes=3):
-		# 	return Response({'result': 'code has expired'})
-
-		# if form_data['code'] == otp.code:
-		# 	user_instance = User.objects.get(pk=pk)
-		#	user_instance.online = True
-		#	user_instance.save()
-		# 	response = Response({'result': 'success'})
-		# 	token = RefreshToken.for_user(user_instance)
-		# 	response.set_cookie('jwt', str(token.access_token), httponly=True)
-		# 	response.set_cookie('refresh', str(token), httponly=True)
-		# 	response.delete_cookie('pk')
-		# 	otp.delete()
-		# 	return response
-		# else:
-		# 	return Response({'result': 'failed'})
 		try:
+			now = timezone.now()
 			pk = request.COOKIES.get('pk')
 			otp = OTPModel.objects.get(user=pk)
-			user_instance = User.objects.get(pk=pk)
-			user_instance.online = True
-			user_instance.save()
-			response = Response({'result': 'success'})
-			token = RefreshToken.for_user(user_instance)
-			response.set_cookie('jwt', str(token.access_token), httponly=True)
-			response.set_cookie('refresh', str(token), httponly=True)
-			response.delete_cookie('pk')
-			otp.delete()
+	
+			form_data = json.loads(request.body)
+
+			if now - otp.created_at > timedelta(minutes=3):
+				return Response({'result': 'code has expired'})
+
+			if form_data['code'] == otp.code:
+				user_instance = User.objects.get(pk=pk)
+				user_instance.online = True
+				user_instance.save()
+				response = Response({'result': 'success'})
+				token = RefreshToken.for_user(user_instance)
+				response.set_cookie('jwt', str(token.access_token), httponly=True)
+				response.set_cookie('refresh', str(token), httponly=True)
+				response.delete_cookie('pk')
+				otp.delete()
+				return response
+			else:
+				return Response({'result': 'failed'})
 		except:
 			return Response(status=status.HTTP_400_BAD_REQUEST)
